refactor(ymax-planner): Export and test `scaleToNat`
